Moved common methods between Adhesion and CardAdhesion into a common AbstractAdhesion class for better handling of runtime adhesion type checkings
