+++ src/Models/Adhesion.php
@@ -2,47 +2,8 @@
 
 namespace Pagos360\Models;
 
-class Adhesion extends AbstractModel
+class Adhesion extends AbstractAdhesion
 {
-    /**
-     * ID de Adhesión.
-     *
-     * @var int
-     */
-    protected $id;
-
-    /**
-     * Estado de la Adhesión.
-     * Los posibles valores son: "pending_to_sign", "signed", "canceled".
-     *
-     * @var string
-     */
-    protected $state;
-
-    /**
-     * Fecha y hora de creación.
-     *
-     * @var \DateTimeInterface
-     */
-    protected $createdAt;
-
-    /**
-     * Nombre del titular del servicio que se debitará.
-     *
-     * @var string
-     */
-    protected $adhesionHolderName;
-
-    /**
-     * Este atributo se puede utilizar como referencia para identificar la
-     * Adhesión y sincronizar con tus sistemas de backend el origen de
-     * la operación. Algunos valores comunmente utilizados son: ID de Cliente,
-     * DNI, CUIT, ID de venta o Nro. de Factura entre otros.
-     *
-     * @var string|null
-     */
-    protected $externalReference;
-
     /**
      * Número de CBU de la cuenta bancaria en la que se ejecutarán los débitos.
      *
@@ -65,25 +26,11 @@ class Adhesion extends AbstractModel
     protected $cbuHolderName;
 
     /**
-     * Email del del titular de la cuenta bancaria.
-     *
-     * @var string
-     */
-    protected $email;
-
-    /**
      * Nombre de la entidad bancaria a la que corresponde el número de CBU.
      *
      * @var string
      */
     protected $bank;
-
-    /**
-     * Descripción o concepto de la Adhesión.
-     *
-     * @var string
-     */
-    protected $description;
 
     /**
      * Descripción Bancaria que se mostrará en el resumen de la cuenta bancaria
@@ -92,55 +39,6 @@ class Adhesion extends AbstractModel
      * @var string
      */
     protected $shortDescription;
-
-    /**
-     * Fecha y hora de cancelación. Si está presente, este valor
-     * indica la fecha en que la adhesion ha sido cancelada.
-     *
-     * @var \DateTimeInterface|null
-     */
-    protected $canceledAt;
-
-    /**
-     * Motivo de cancelación de una Adhesión. Si está presente, este valor
-     * indica por qué la adhesion ha sido cancelada.
-     *
-     * @var string|null
-     */
-    protected $stateComment;
-
-    /**
-     * Objeto JSON que se puede utilizar para guardar atributos adicionales en
-     * la adhesion y poder sincronizar con tus sistemas de backend. Pagos360 no
-     * utiliza este objeto.
-     *
-     * @var array|null
-     */
-    protected $metadata;
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getState(): string
-    {
-        return $this->state;
-    }
-
-    /**
-     * @return \DateTimeInterface
-     */
-    public function getCreatedAt(): \DateTimeInterface
-    {
-        return $this->createdAt;
-    }
 
     /**
      * @return string
@@ -157,24 +55,6 @@ class Adhesion extends AbstractModel
     public function setAdhesionHolderName(string $adhesionHolderName): self
     {
         $this->adhesionHolderName = $adhesionHolderName;
-        return $this;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getExternalReference(): ?string
-    {
-        return $this->externalReference;
-    }
-
-    /**
-     * @param string|null $externalReference
-     * @return self
-     */
-    public function setExternalReference(?string $externalReference): self
-    {
-        $this->externalReference = $externalReference;
         return $this;
     }
 
@@ -235,45 +115,9 @@ class Adhesion extends AbstractModel
     /**
      * @return string
      */
-    public function getEmail(): string
-    {
-        return $this->email;
-    }
-
-    /**
-     * @param string $email
-     * @return self
-     */
-    public function setEmail(string $email): self
-    {
-        $this->email = $email;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
     public function getBank(): string
     {
         return $this->bank;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDescription(): string
-    {
-        return $this->description;
-    }
-
-    /**
-     * @param string $description
-     * @return self
-     */
-    public function setDescription(string $description): self
-    {
-        $this->description = $description;
-        return $this;
     }
 
     /**
@@ -294,13 +138,6 @@ class Adhesion extends AbstractModel
         return $this;
     }
 
-    /**
-     * @return \DateTimeInterface|null
-     */
-    public function getCanceledAt(): ?\DateTimeInterface
-    {
-        return $this->canceledAt;
-    }
 
     /**
      * @return string|null
@@ -310,21 +147,5 @@ class Adhesion extends AbstractModel
         return $this->stateComment;
     }
 
-    /**
-     * @return array|null
-     */
-    public function getMetadata(): ?array
-    {
-        return $this->metadata;
-    }
 
-    /**
-     * @param array|null $metadata
-     * @return self
-     */
-    public function setMetadata(?array $metadata): self
-    {
-        $this->metadata = $metadata;
-        return $this;
-    }
 }
