+++ src/Models/CardAdhesion.php
@@ -2,39 +2,8 @@
 
 namespace Pagos360\Models;
 
-class CardAdhesion extends AbstractModel
+class CardAdhesion extends AbstractAdhesion
 {
-    /**
-     * Nombre del titular del servicio que se debitará.
-     *
-     * @var string
-     */
-    protected $adhesionHolderName;
-
-    /**
-     * Email del titular de la Tarjeta.
-     *
-     * @var string
-     */
-    protected $email;
-
-    /**
-     * Descripción o concepto de la Adhesión.
-     *
-     * @var string
-     */
-    protected $description;
-
-    /**
-     * Este atributo se puede utilizar como referencia para identificar la
-     * Adhesión y sincronizar con tus sistemas de backend el origen de la
-     * operación. Algunos valores comúnmente utilizados son: ID de Cliente, DNI,
-     * CUIT, ID de venta o Nro. de Factura entre otros.
-     *
-     * @var string
-     */
-    protected $externalReference;
-
     /**
      * Hash en Base64 que contiene la Encriptación del Número de Tarjeta en la
      * que se ejecutarán los débitos automáticos.
@@ -51,22 +20,6 @@ class CardAdhesion extends AbstractModel
     protected $cardHolderName;
 
     /**
-     * Objeto JSON que se puede utilizar para guardar atributos adicionales en
-     * la adhesión y poder sincronizar con tus sistemas de backend.
-     * Pagos360.com no utiliza este objeto.
-     *
-     * @var array|null
-     */
-    protected $metadata;
-
-    /**
-     * ID de Adhesión.
-     *
-     * @var int
-     */
-    protected $id;
-
-    /**
      * Ultimos 4 numeros de la Tarjeta.
      *
      * @var string
@@ -79,36 +32,6 @@ class CardAdhesion extends AbstractModel
      * @var string
      */
     protected $card;
-
-    /**
-     * Estado de la Adhesión.
-     * Los posibles valores son: "pending_to_sign", "signed", "canceled".
-     *
-     * @var string
-     */
-    protected $state;
-
-    /**
-     * Fecha y hora de creación.
-     *
-     * @var \DateTimeImmutable
-     */
-    protected $createdAt;
-
-    /**
-     * Motivo de cancelación de una Adhesión.
-     *
-     * @var string|null
-     */
-    protected $stateComment;
-
-    /**
-     * Fecha y hora de cancelación. Si está presente, este valor
-     * indica la fecha en que la adhesion ha sido cancelada.
-     *
-     * @var \DateTimeImmutable|null
-     */
-    protected $canceledAt;
 
     /**
      * @return string
@@ -125,60 +48,6 @@ class CardAdhesion extends AbstractModel
     public function setAdhesionHolderName(string $adhesionHolderName): CardAdhesion
     {
         $this->adhesionHolderName = $adhesionHolderName;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getEmail(): string
-    {
-        return $this->email;
-    }
-
-    /**
-     * @param string $email
-     * @return CardAdhesion
-     */
-    public function setEmail(string $email): CardAdhesion
-    {
-        $this->email = $email;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDescription(): string
-    {
-        return $this->description;
-    }
-
-    /**
-     * @param string $description
-     * @return CardAdhesion
-     */
-    public function setDescription(string $description): CardAdhesion
-    {
-        $this->description = $description;
-        return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getExternalReference(): string
-    {
-        return $this->externalReference;
-    }
-
-    /**
-     * @param string $externalReference
-     * @return CardAdhesion
-     */
-    public function setExternalReference(string $externalReference): CardAdhesion
-    {
-        $this->externalReference = $externalReference;
         return $this;
     }
 
@@ -211,32 +80,6 @@ class CardAdhesion extends AbstractModel
     }
 
     /**
-     * @return array|null
-     */
-    public function getMetadata(): ?array
-    {
-        return $this->metadata;
-    }
-
-    /**
-     * @param array $metadata
-     * @return CardAdhesion
-     */
-    public function setMetadata(array $metadata): CardAdhesion
-    {
-        $this->metadata = $metadata;
-        return $this;
-    }
-
-    /**
-     * @return int
-     */
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    /**
      * @return string
      */
     public function getLastFourDigits(): string
@@ -253,22 +96,6 @@ class CardAdhesion extends AbstractModel
     }
 
     /**
-     * @return string
-     */
-    public function getState(): string
-    {
-        return $this->state;
-    }
-
-    /**
-     * @return \DateTimeImmutable
-     */
-    public function getCreatedAt(): \DateTimeImmutable
-    {
-        return $this->createdAt;
-    }
-
-    /**
      * @return string|null
      */
     public function getStateComment(): ?string
@@ -276,11 +103,4 @@ class CardAdhesion extends AbstractModel
         return $this->stateComment;
     }
 
-    /**
-     * @return \DateTimeImmutable|null
-     */
-    public function getCanceledAt(): ?\DateTimeImmutable
-    {
-        return $this->canceledAt;
-    }
 }
